requirements.txt
annotated-types==0.7.0    
anyio==4.9.0    
asttokens==3.0.0    
certifi==2025.6.15
click==8.2.1    
comm==0.2.2    
coverage==7.9.2    
debugpy==1.8.12   
decorator==5.1.1    
distlib==0.3.9    
dnspython==2.7.0    
email_validator==2.2.0    
executing==2.2.0    
fastapi==0.115.14
fastapi-cli==0.0.7
filelock==3.17.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.3.1
onnx==1.18.0
packaging==24.2
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.6.0
prompt_toolkit==3.0.50
protobuf==6.31.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.1
rich==14.0.0
rich-toolkit==0.14.8
scikit-learn==1.7.0
scipy==1.16.0
shellingham==1.5.4
six==1.17.0
skl2onnx==1.19.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
threadpoolctl==3.6.0
tornado==6.4.2
traitlets==5.14.3
typer==0.16.0
typing_extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
firebase-admin==6.9.0